requirements.txt
filelock==3.25.2
fsspec==2026.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
opt_einsum==3.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==3.0.2
pyro-api==0.1.2
pyro-ppl==1.9.1
pysam==0.23.3
python-dateutil==2.9.0.post0
scikit-misc==0.5.1
six==1.17.0
sympy==1.14.0
torch==2.11.0
tqdm==4.67.3
typing_extensions==4.15.0

